Unify `FeedViewController` static factory helper into the `FeedUIComposer`
